video upload routes and controller added, HLS conversion pending

+++ src/controllers/channel.controller.js
@@ -11,14 +11,14 @@ import fs from "fs/promises"
 const createChannel = async(req,res)=>{
     const user = await Users.findById(req.userId);
     if(!user){
-        return res.status(404).send(new ApiResponse(404, "User Does Not exsits"))
+        return res.status(404).send(new ApiResponse(404, "User Does Not exist"))
     }   
 
 
     const channelUserName = req.body.channelUserName;
     // const channel = await Channels.findOne({user_id:user._id});
     if(user.userType==2){
-        return res.status(400).send(new ApiResponse(400, "channel Already exsits"))
+        return res.status(400).send(new ApiResponse(400, "channel Already exist"))
         
     }
     else{
@@ -30,13 +30,18 @@ const createChannel = async(req,res)=>{
     
     
     try{
+        let filePath = "";
+        if(req.fileName){
+            filePath = `${env.UPLOAD_COVER_IMAGE_FOLDER}/${req.fileName[0]?.name}`
+        }
 
         const dataToSave = {
         "channelName" : req.body.channelName,
         "user_id" : user._id ,
         "description" :req.body.description,
         "channelUserName" : channelUserName,
-        "coverImage" : `${env.UPLOAD_COVER_IMAGE_FOLDER}/${req.fileName?.name}`,
+        "profilePhoto" : user.profilePhoto,
+        "coverImage" : filePath,
         "contactInfo" : req.body.contactInfo,
         "homeTabSetting" :{'sortBy' : req.body.homeTabSetting?.sortBy},
         "totalSubscriberCount" : 0 ,
@@ -106,10 +111,14 @@ const updateChannelCoverImage  = async(req,res)=>{
     if(!channel){
         return res.status(404).send(new ApiResponse(404, "User does not have a channel"));
     }
+    if(!req.fileName){
+        return res.status(400).send(new ApiResponse(400,"Please upload a new cover Image"))
+    }
 
     try{
+        
         const oldCoverImageFilePath = channel.coverImage;
-         await Channels.findByIdAndUpdate({_id:channel._id},{$set : {"coverImage" : `${env.UPLOAD_COVER_IMAGE_FOLDER}/${req.fileName?.name}`}});
+         await Channels.findByIdAndUpdate({_id:channel._id},{$set : {"coverImage" : `${env.UPLOAD_COVER_IMAGE_FOLDER}/${req.fileName[0]?.name}`}});
         try{
             await fs.unlink(oldCoverImageFilePath)
         }catch(err){
@@ -124,7 +133,27 @@ const updateChannelCoverImage  = async(req,res)=>{
     }
 }
 
+const getChannelDetails = async(req,res)=>{
+    // console.log(req.params.userName);
+    // const userId = new mongoose.Types.ObjectId(req.userId)
+
+    const channel = await Channels.findOne({channelUserName:String(req.params.userName)}).select("-user_id -_id");
+    if(!channel){
+        return res.status(404).send(new ApiResponse(404, "Channel Does Not exist"));
+    }
+    const data = {
+        "channelName" : channel.channelName ,
+        "description" : channel.description ,
+        "channelUserName" : channel.channelUserName ,
+        "profilePhoto" : channel.profilePhoto ,
+        "coverImage" : channel.coverImage ,
+        "contactInfo" : channel.contactInfo ,
+        "homeTabSetting" : channel.homeTabSetting ,
+        "totalSubscriberCount" : channel.totalSubscriberCount ,
+        "totalViewCount" : channel.totalViewCount 
+    }
+    return res.status(200).send(new ApiResponse(200, "Channel Details" , data ))
+}
 
 
-
-export {createChannel, updateChannelDetails, updateChannelCoverImage}
+export {createChannel, updateChannelDetails, updateChannelCoverImage , getChannelDetails}
